requirements.txt
django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
